requirements.txt
pika
python-dotenv
psycopg2-binary
openai
google-genai
boto3